labs/3-Deploy-App-to-ACR/app/requirements.txt
Flask==3.1.3
Werkzeug==3.1.8
azure-monitor-opentelemetry==1.8.7
